+++ components/PostsList/PostsList.js
@@ -1,5 +1,5 @@
 import classes from "./postsList.module.scss";
-import Post from "../Post/Post";
+import Feed from "../Feed/Feed";
 import Link from '../Link/Link';
 import React from "react";
 
@@ -7,7 +7,7 @@ export default function PostsList({posts}) {
     return (
         <div className={classes.container}>
             {posts.map((elem, idx) => (
-                <Post key={idx} {...elem} />
+                <Feed key={idx} {...elem} />
             ))}
         </div>
     )
