+++ package.json
@@ -21,11 +21,13 @@
   "dependencies": {
     "@draft-js-plugins/editor": "^4.1.3",
     "@draft-js-plugins/image": "^4.1.3",
+    "@reduxjs/toolkit": "^1.8.5",
     "draft-js": "^0.11.7",
     "firebase": "^9.9.0",
     "next": "^12.2.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
     "sass": "^1.53.0"
   }
 }
