+++ pages/_app.js
@@ -1,6 +1,12 @@
 import '../styles/global.css';
 import '../styles/EditorStyling.css';
+import {Provider} from "react-redux";
+import store from "../app/store";
 
 export default function App({ Component, pageProps }) {
-    return <Component {...pageProps} />;
+    return (
+        <Provider store={store}>
+            <Component {...pageProps} />
+        </Provider>
+    );
 }
