+++ pages/post/[url].js
@@ -1,20 +1,24 @@
-import postsController from "../../db/controllers/posts.controller";
-import Layout from "../../components/Layout/Layout";
-import PostDetails from "../../components/PostDetails/PostDetails";
-import {useRouter} from "next/router";
 import {useEffect, useState} from "react";
+import {useDispatch, useSelector} from "react-redux";
+import {useRouter} from "next/router";
+
+import postsController from "../../db/controllers/posts.controller";
+
+import Layout from "../../components/Layout/Layout";
+import Post from "../../features/Post";
 import Loading from "../../components/Loading/Loading";
 import Header from "../../components/Header/Header";
 import Main from "../../components/Main/Main";
-import Search from "../../components/Search/Search";
-import Button from "../../components/Button/Button";
 import Link from "../../components/Link/Link";
+import {setPost} from "../../features/postSlice";
 
-export default function Post() {
+export default function PostDetails() {
     const router = useRouter();
     const url = router.query.url;
     const [isFetch, changeIsFetch] = useState(false);
-    const [data, setData] = useState(Object.create(null));
+    const data = useSelector(state => state.post.data);
+    const dispatch = useDispatch();
+    const setPostDispatched = (val) => dispatch(setPost(val));
 
     function deletePost() {
         postsController.deleteById(data.id).then(() => router.push('/'));
@@ -24,11 +28,12 @@ export default function Post() {
         changeIsFetch(true);
         if (url) {
             postsController.getOne(router.query.url)
-                .then((res) => setData(res))
+                .then((res) => setPostDispatched(res))
                 .then(() => changeIsFetch(false))
         }
     }, [url])
 
+    // console.log(data)
     return (
         <Layout title={"Sog"}>
             <Header search>
@@ -36,7 +41,7 @@ export default function Post() {
             </Header>
             <Main>
                 <Main.Section>
-                    { isFetch ? <Loading /> : <PostDetails {...data} deletePost={deletePost}/>}
+                    { isFetch ? <Loading /> : <Post {...data} deletePost={deletePost}/>}
                 </Main.Section>
             </Main>
         </Layout>
